pyproject.toml
[project]
name = "img2latex-vlm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.10.1",
    "datasets>=4.1.1",
    "ipywidgets>=8.1.7",
    "notebook>=7.4.7",
    "peft>=0.17.1",
    "pytest>=8.4.2",
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "tqdm>=4.67.1",
    "transformers>=4.46.2",
    "wandb>=0.18.1",
    "google-cloud-aiplatform>=1.38.0",
]
